rename pkg to vendor, pack ftml 0.8.1
