+++ client/.prettierignore
@@ -1,6 +1,6 @@
 dist
 node_modules
-pkg
+vendor
 
 # File types we don't want prettier to care about, but normally does
 *.json
