+++ client/modules/ftml-wasm/package.json
@@ -9,7 +9,7 @@
   "scripts": {
     "clean": "trash dist .ultra.cache.json",
     "build": "node ../../scripts/build-modules.js -self",
-    "pack-ftml": "cd ../../../ftml && wasm-pack build --release --target web --out-dir ../client/modules/ftml-wasm/pkg && cd ../client/modules/ftml-wasm/pkg && trash .gitignore LICENSE.md package.json README.md"
+    "pack-ftml": "cd ../../../ftml && wasm-pack build --release --target web --out-dir ../client/modules/ftml-wasm/vendor -- --no-default-features && cd ../client/modules/ftml-wasm/vendor && trash .gitignore LICENSE.md package.json README.md"
   },
   "type": "module",
   "main": "dist/index.js",
